backend/requirements.txt
fastapi >= 0.95
uvicorn[standard]>=0.22
httpx>=0.24
sqlalchemy>=2.0
python-chess>=1.999